+++ kipper/core/src/compiler/semantics/const.ts
@@ -199,6 +199,20 @@ export type KipperReturnType = KipperPrimitiveType | KipperListType<any>;
 export const kipperReturnTypes: Array<string> = [...kipperPrimitiveTypes, "list"];
 
 /**
+ * List of all supported variable type conversions that can be performed in a Kipper program.
+ *
+ * For each translation, there will have to be a corresponding {@link KipperTargetBuiltInGenerator generator function},
+ * which generates for each conversion the translator function in the specific target.
+ * @since 0.8.0
+ */
+export const kipperSupportedConversions: Array<[KipperType, KipperType]> = [
+  ["num", "str"],
+  ["str", "num"],
+  ["bool", "str"],
+  ["bool", "num"],
+];
+
+/**
  * All available storage types inside Kipper.
  */
 export type KipperStorageType = "var" | "const";